uppstädat och sorterat efter mergekonf med dubbla filer
