+++ frontend/src/hooks/index.ts
@@ -1,5 +1,5 @@
 export { useWeather } from "./useWeather";
 export { useFlightInfo } from "./useFlightInfo";
-export { useMessages } from "./useMessages";
+export { useMessages } from "./useMessagesAdmin";
 export { useDepartureFlights } from "./useDepartureFlights";
 export { useArrivalFlights } from "./useArrivalFlights";
